Chore : rename (entity -> model)
