+++ data/src/main/java/com/teamyg/data/repository/KakaoUserRepositoryImpl.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import com.kakao.sdk.common.model.ClientError
 import com.kakao.sdk.common.model.ClientErrorCause
 import com.kakao.sdk.user.UserApiClient
-import com.teamyg.domain.entity.KakaoLoginResult
+import com.teamyg.domain.model.KakaoLoginResult
 import com.teamyg.domain.repository.KakaoUserRepository
 import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.coroutines.ExperimentalCoroutinesApi
