+++ feature/login/impl/src/main/java/com/teamyg/login/impl/LoginViewModel.kt
@@ -2,7 +2,7 @@ package com.teamyg.login.impl
 
 import android.util.Log
 import androidx.lifecycle.viewModelScope
-import com.teamyg.domain.entity.KakaoLoginResult
+import com.teamyg.domain.model.KakaoLoginResult
 import com.teamyg.domain.usecase.LoginWithKakaoUseCase
 import com.tjyg.core.ui.BaseViewModel
 import com.tjyg.core.ui.UiIntent
